+++ packages/wichita/test/script-test.js
@@ -26,6 +26,19 @@ describe('Script', () => {
 			assert.strictEqual(dom.window.document.title, 'initial value, update from script');
 		});
 
+		it('evaluates script with exports', async function () {
+			const dom = new JSDOM('<title>with exports</title>');
+			const script = new Script(getTestPath(this), `
+				export default document.title + '?';
+				export const named = document.title + '!';
+			`);
+
+			const exports = await script.evaluate(dom.window);
+
+			assert.strictEqual(exports.default, 'with exports?');
+			assert.strictEqual(exports.named, 'with exports!');
+		});
+
 		it('evaluates script with imports', async function () {
 			const dom = new JSDOM('<title>initial value</title>');
 			const script = new Script(getTestPath(this), `
@@ -40,17 +53,17 @@ describe('Script', () => {
 			assert.strictEqual(dom.window.document.title, 'initial value, update from source component, update from package component');
 		});
 
-		it('evaluates script with exports', async function () {
-			const dom = new JSDOM('<title>with exports</title>');
+		it('evaluates script with non-js imports', async function () {
+			const dom = new JSDOM('');
 			const script = new Script(getTestPath(this), `
-				export default document.title + '?';
-				export const named = document.title + '!';
+				export { default as json } from './files/non-js-dependency.json' with { type: 'json' };
+				export { default as text } from './files/non-js-dependency.md' with { type: 'text' };
 			`);
 
 			const exports = await script.evaluate(dom.window);
 
-			assert.strictEqual(exports.default, 'with exports?');
-			assert.strictEqual(exports.named, 'with exports!');
+			assert.deepEqual(exports.json, { id: 'json-dependency' });
+			assert.strictEqual(exports.text, 'dependency of type `"text"`\n');
 		});
 
 		it('evaluates script multiple times', async function () {
@@ -62,14 +75,74 @@ describe('Script', () => {
 				export const times = ++i;
 			`);
 
-			const exports = await Promise.all([
-				script.evaluate(dom1.window),
-				script.evaluate(dom2.window),
-			]);
+			const exports1 = await script.evaluate(dom1.window);
+			const exports2 = await script.evaluate(dom2.window);
 
-			assert.strictEqual(exports[0].default, 'once!');
-			assert.strictEqual(exports[1].default, 'twice!');
-			assert.strictEqual(exports[1].times, 1);
+			assert.strictEqual(exports1.default, 'once!');
+			assert.strictEqual(exports2.default, 'twice!');
+			assert.strictEqual(exports2.times, 1);
+		});
+	});
+
+	describe('.resolve()', () => {
+		it('enables custom resolving', async function () {
+			class ScriptWithCustomResolver extends Script {
+				resolve (specifier, parentIdentifier) {
+					if (specifier !== 'dependency')
+						return super.resolve(specifier, parentIdentifier);
+
+					return path.resolve('./test/files/mock-dependency.mjs');
+				}
+			}
+
+			const dom = new JSDOM('<title>initial value</title>');
+			const script = new ScriptWithCustomResolver(getTestPath(this), `
+				import dependency from 'dependency';
+				import update from './files/helpers/update.mjs';
+
+				update(dependency);
+			`);
+
+			await script.evaluate(dom.window);
+
+			assert.strictEqual(dom.window.document.title, 'initial value, update from mock dependency');
+		});
+	});
+
+	describe('.load()', () => {
+		it('enables custom loading', async function () {
+			class ScriptWithCustomLoader extends Script {
+				load (identifier) {
+					if (!identifier.endsWith('custom-component.mjs'))
+						return super.load(identifier);
+
+					return Promise.resolve(`
+						import update from '../helpers/update.mjs';
+
+						export default function component () {
+							update('custom component');
+						}
+					`);
+				}
+			}
+
+			const dom = new JSDOM('<title>initial value</title>');
+			const script = new ScriptWithCustomLoader(getTestPath(this), `
+				import update, { memory as _helperMemory } from './files/helpers/update.mjs';
+				import component from './files/source-component/source-component.mjs';
+				import customComponent from './files/custom-component/custom-component.mjs';
+
+				update('entry');
+				component();
+				customComponent();
+
+				export { _helperMemory };
+			`);
+
+			const { _helperMemory } = await script.evaluate(dom.window);
+
+			assert.strictEqual(dom.window.document.title, 'initial value, update from entry, update from source component, update from custom component');
+			assert.deepEqual(_helperMemory, [ 'update from entry', 'update from source component', 'update from custom component' ]);
 		});
 	});
 });